# chore: Improve startup times on Android through Swaps changes