--- a/app/core/Engine.ts
+++ b/app/core/Engine.ts
@@ -94,7 +94,10 @@ import {
   SubjectMetadataControllerState,
   ///: END:ONLY_INCLUDE_IF
 } from '@metamask/permission-controller';
-import SwapsController, { swapsUtils } from '@metamask/swaps-controller';
+import SwapsController, {
+  swapsUtils,
+  SwapsControllerState,
+} from '@metamask/swaps-controller';
 import {
   PPOMController,
   PPOMControllerEvents,
@@ -187,8 +190,6 @@ import {
   SignatureControllerEvents,
 } from '@metamask/signature-controller';
 import { hasProperty, Hex, Json } from '@metamask/utils';
-// TODO: Export this type from the package directly
-import { SwapsState } from '@metamask/swaps-controller/dist/SwapsController';
 import { providerErrors } from '@metamask/rpc-errors';
 
 import { PPOM, ppomInit } from '../lib/ppom/PPOMView';
@@ -318,7 +319,7 @@ export interface EngineState {
   TokenRatesController: TokenRatesState;
   TransactionController: TransactionControllerState;
   SmartTransactionsController: SmartTransactionsControllerState;
-  SwapsController: SwapsState;
+  SwapsController: SwapsControllerState;
   GasFeeController: GasFeeState;
   TokensController: TokensState;
   TokenDetectionController: BaseState;
@@ -1366,29 +1367,42 @@ class Engine {
       this.smartTransactionsController,
       new SwapsController(
         {
+          messenger: this.controllerMessenger.getRestricted({
+            name: 'SwapsController',
+            // TODO: allow these internal calls once GasFeeController
+            // export these action types and register its action handlers
+            // allowedActions: [
+            //   'GasFeeController:getEIP1559GasFeeEstimates',
+            // ],
+            allowedActions: [],
+            allowedEvents: [],
+          }),
+          // TODO: Remove once GasFeeController exports this action type
           fetchGasFeeEstimates: () => gasFeeController.fetchGasFeeEstimates(),
           // @ts-expect-error TODO: Resolve mismatch between gas fee and swaps controller types
           fetchEstimatedMultiLayerL1Fee,
         },
         {
-          clientId: AppConstants.SWAPS.CLIENT_ID,
-          fetchAggregatorMetadataThreshold:
-            AppConstants.SWAPS.CACHE_AGGREGATOR_METADATA_THRESHOLD,
-          fetchTokensThreshold: AppConstants.SWAPS.CACHE_TOKENS_THRESHOLD,
-          fetchTopAssetsThreshold:
-            AppConstants.SWAPS.CACHE_TOP_ASSETS_THRESHOLD,
-          supportedChainIds: [
-            swapsUtils.ETH_CHAIN_ID,
-            swapsUtils.BSC_CHAIN_ID,
-            swapsUtils.SWAPS_TESTNET_CHAIN_ID,
-            swapsUtils.POLYGON_CHAIN_ID,
-            swapsUtils.AVALANCHE_CHAIN_ID,
-            swapsUtils.ARBITRUM_CHAIN_ID,
-            swapsUtils.OPTIMISM_CHAIN_ID,
-            swapsUtils.ZKSYNC_ERA_CHAIN_ID,
-            swapsUtils.LINEA_CHAIN_ID,
-            swapsUtils.BASE_CHAIN_ID,
-          ],
+          config: {
+            clientId: AppConstants.SWAPS.CLIENT_ID,
+            fetchAggregatorMetadataThreshold:
+              AppConstants.SWAPS.CACHE_AGGREGATOR_METADATA_THRESHOLD,
+            fetchTokensThreshold: AppConstants.SWAPS.CACHE_TOKENS_THRESHOLD,
+            fetchTopAssetsThreshold:
+              AppConstants.SWAPS.CACHE_TOP_ASSETS_THRESHOLD,
+            supportedChainIds: [
+              swapsUtils.ETH_CHAIN_ID,
+              swapsUtils.BSC_CHAIN_ID,
+              swapsUtils.SWAPS_TESTNET_CHAIN_ID,
+              swapsUtils.POLYGON_CHAIN_ID,
+              swapsUtils.AVALANCHE_CHAIN_ID,
+              swapsUtils.ARBITRUM_CHAIN_ID,
+              swapsUtils.OPTIMISM_CHAIN_ID,
+              swapsUtils.ZKSYNC_ERA_CHAIN_ID,
+              swapsUtils.LINEA_CHAIN_ID,
+              swapsUtils.BASE_CHAIN_ID,
+            ],
+          },
         },
       ),
       gasFeeController,