--- a/package.json
+++ b/package.json
@@ -190,7 +190,7 @@
     "@metamask/snaps-sdk": "^6.2.1",
     "@metamask/snaps-utils": "^8.0.1",
     "@metamask/swappable-obj-proxy": "^2.1.0",
-    "@metamask/swaps-controller": "^9.0.2",
+    "@metamask/swaps-controller": "^10.0.0",
     "@metamask/transaction-controller": "^35.0.0",
     "@metamask/utils": "^8.1.0",
     "@ngraveio/bc-ur": "^1.1.6",