<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been completely filled out, and PR status checks have passed at least once.
-->

## **Description**

<!--
Write a short description of the changes included in this pull request, also include relevant motivation and context. Have in mind the following questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

This PR removes Web3 in favor of just using Ethers to read values off a contract.

This has resulted in the following performance improvements:

|           | With Web3 | No Web3 | Perf improvement |
|-----------|-----------|---------|------------------|
| App Start | 4920ms    | 658ms   | ~86% decrease             |
| Logcat    | 6929ms    | ~2500ms | ~63% decrease             |

App start and Logcat are 2 related but different measurements. Logcat is more closely related to visually looking at the app and seeing when it starts up.


## **Related issues**

Fixes: https://github.com/MetaMask/metamask-mobile/issues/10611
Related to: https://github.com/MetaMask/swaps-controller/pull/293

## **Manual testing steps**

You will need to run the QA build of the Android app onto a physical device to measure performance changes accurately.

1. `android/app/build.gradle`
    1. Comment out line 3: `apply plugin: "io.sentry.android.gradle"`
    1. Comment out line 62 to 100 to get rid of Sentry (starts with `// Sentry environment variables used by Sentry CLI to upload files. Upload is disabled by default`)
    1. Line 242: change to `signingConfigs.debug`
1. Connect a physical Android device to your machine.
1. Run this command to build and install onto your physical device: `cd ./android && ./gradlew assembleQaRelease -PminSdkVersion=26 -DtestBuildType=release -x app:createBundleFlaskDebugJsAndAssets && cd ../ && adb install -r android/app/build/outputs/apk/qa/release/app-qa-release.apk `


### Testing

1. Start the Android app
3. Notice how much faster it got
4. Go to Swaps
5. Select From token as an ERC20 which requires allowance
6. Get a quote
7. Notice Edit Limit
8. 5. Select From token as an ERC20 which doesn't requires allowance
6. Get a quote
7. Notice Edit Limit is not there

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Contributor Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile Coding Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format if applicable
- [x] I’ve applied the right labels on the PR (see [labeling guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)). Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described in the ticket it closes and includes the necessary testing evidence such as recordings and or screenshots.